CP-54351 Move trim_util from /opt
