+++ Makefile
@@ -41,6 +41,7 @@ SM_LIBS += refcounter
 SM_LIBS += journaler
 SM_LIBS += fjournaler
 SM_LIBS += pluginutil
+SM_LIBS += trim_util
 
 # Things used as commands which install in libexec
 # which are in python and need compatibility symlinks from
